// src/interfaces/interfacesFormValues.tsx
export interface InterfaceFormValues {
    name: string;
    type?: string;
    age?: number | undefined;
    weight?: number | undefined;
    sex: string;
    breed: string;
    size: string;
    vaccinated: boolean;
    deworming: boolean;
    photo: string;
    sterilized: boolean;
    email: string;
    phone?: string | undefined;
    cell?: string | undefined;
    description: string;
    available: boolean;
  }